Django/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.26
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0